backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
beautifulsoup4==4.12.3
pydantic==2.10.3
google-generativeai
python-dotenv==1.0.1
lxml==5.3.0
